move poison message handler to superclass
