+++ src/DurableTask.AzureStorage/Messaging/ControlQueue.cs
@@ -50,36 +50,6 @@ namespace DurableTask.AzureStorage.Messaging
 
         protected override TimeSpan MessageVisibilityTimeout => this.settings.ControlQueueVisibilityTimeout;
 
-        private async Task HandleIfPoisonMessageAsync(MessageData messageData)
-        {
-            var isPoison = false;
-            var queueMessage = messageData.OriginalQueueMessage;
-
-            if (queueMessage.DequeueCount > this.settings.PoisonMessageDeuqueCountThreshold)
-            {
-                var poisonMessage = new DynamicTableEntity(queueMessage.Id, this.Name)
-                {
-                    Properties =
-                    {
-                        ["RawMessage"] = new EntityProperty(queueMessage.Message)
-                    }
-                };
-
-                // add to poison table
-                var poisonMessageTableName = this.settings.TaskHubName.ToLowerInvariant() + "-poison";
-                var poisonMessagesTable = this.azureStorageClient.GetTableReference(poisonMessageTableName);
-                await poisonMessagesTable.CreateIfNotExistsAsync();
-                await poisonMessagesTable.InsertAsync(poisonMessage);
-
-                // delete from queue so it doesn't get processed again.
-                await this.storageQueue.DeleteMessageAsync(queueMessage);
-
-                // since isPoison is `true`, we'll override the deserialized message w/ a suspend event
-                isPoison = true;
-            }
-            messageData.TaskMessage.Event.IsPoison = isPoison;
-        }
-
         public async Task<IReadOnlyList<MessageData>> GetMessagesAsync(CancellationToken cancellationToken)
         {
             using (var linkedCts = CancellationTokenSource.CreateLinkedTokenSource(this.releaseCancellationToken, cancellationToken))
